Split target-first chained search tests
Move the specialized regression coverage into a focused partial test class and cover a following source predicate.
